+++ internal/cli/root.go
@@ -33,6 +33,7 @@ func NewRootCommand() *cobra.Command {
 	root.AddCommand(templateCmd(&dataDir))
 	root.AddCommand(apiCmd(&dataDir))
 	root.AddCommand(recordCmd(&dataDir))
+	root.AddCommand(sensorCmd(&dataDir))
 	root.AddCommand(observeCmd(&dataDir, &daemonURL))
 	root.AddCommand(telemetryCmd(&dataDir, &daemonURL))
 	root.AddCommand(timelineCmd(&dataDir, &daemonURL))